refactor(libs): extract geometry solvers, layout services, and WindowRegistry into libraries
Move daemon-internal implementations into their respective shared
libraries so any compositor can use them without linking the daemon:

- enforceWindowMinSizes + removeZoneOverlaps → phosphor-geometry
  (boundary solver + pairwise fallback, min-size-aware multi-pass)
- LayoutFactory → phosphor-zones (type-based layout creation)
- LayoutWorker + LayoutComputeService → phosphor-zones (threaded
  async geometry computation with generation-based coalescing)
- WindowRegistry → phosphor-engine-api (window identity tracking
  with canonicalization and reverse-index lookup)

Daemon retains thin forwarding headers with namespace aliases.
Also fixes a latent missing-include in configmigration.cpp exposed
by unity-batch reshuffling.
