+++ libs/phosphor-zones/src/layoutworker.cpp
@@ -1,11 +1,10 @@
 // SPDX-FileCopyrightText: 2026 fuddlesworth
-// SPDX-License-Identifier: GPL-3.0-or-later
+// SPDX-License-Identifier: LGPL-2.1-or-later
 
-#include "layoutworker.h"
-
+#include <PhosphorZones/LayoutWorker.h>
 #include <PhosphorZones/Zone.h>
 
-namespace PlasmaZones {
+namespace PhosphorZones {
 
 LayoutWorker::LayoutWorker(QObject* parent)
     : QObject(parent)
@@ -24,14 +23,12 @@ void LayoutWorker::computeGeometries(const LayoutSnapshot& snapshot, uint64_t ge
     for (const auto& zone : snapshot.zones) {
         ComputedZoneGeometry computed;
         computed.zoneId = zone.id;
-        // Shared pure helper: same math PhosphorZones::Zone::calculateAbsoluteGeometry uses
-        // on the main thread. Keeps the two paths byte-identical.
-        computed.absoluteGeometry = PhosphorZones::Zone::computeAbsoluteGeometry(
-            zone.geometryMode, zone.relativeGeometry, zone.fixedGeometry, snapshot.screenGeometry);
+        computed.absoluteGeometry = Zone::computeAbsoluteGeometry(zone.geometryMode, zone.relativeGeometry,
+                                                                  zone.fixedGeometry, snapshot.screenGeometry);
         result.zones.append(computed);
     }
 
     Q_EMIT geometriesReady(result);
 }
 
-} // namespace PlasmaZones
+} // namespace PhosphorZones
