+++ src/core/layoutworker/layoutworker.h
@@ -3,30 +3,9 @@
 
 #pragma once
 
-#include "types.h"
-#include <QObject>
+// Forwarding header — LayoutWorker now lives in phosphor-zones.
+#include <PhosphorZones/LayoutWorker.h>
 
 namespace PlasmaZones {
-
-/**
- * @brief Pure-compute worker for layout zone geometry.
- *
- * Lives on a dedicated QThread (moved via moveToThread). Receives
- * immutable LayoutSnapshot inputs, produces LayoutComputeResult
- * outputs. Has no access to any QObject on the main thread.
- */
-class LayoutWorker : public QObject
-{
-    Q_OBJECT
-
-public:
-    explicit LayoutWorker(QObject* parent = nullptr);
-
-public Q_SLOTS:
-    void computeGeometries(const PlasmaZones::LayoutSnapshot& snapshot, uint64_t generation);
-
-Q_SIGNALS:
-    void geometriesReady(const PlasmaZones::LayoutComputeResult& result);
-};
-
+using LayoutWorker = PhosphorZones::LayoutWorker;
 } // namespace PlasmaZones
